mvp_streamlit/requirements.txt
streamlit==1.55.0
chromadb==1.5.5
networkx==3.1
pyvis==0.3.2
sentence-transformers==5.3.0
openai==1.109.1
python-dotenv==1.2.1
